requirements.txt
baostock
pandas
pyarrow
tqdm
psycopg2-binary
duckdb
redis
alphalens
vectorbt
backtrader
numpy
